Se ajustan las clases donde se logra probar toda la funcionalidad. Funciona en MySql, no se logra correr en Postgres.

+++ src/main/java/com/platzi/market/domain/repository/ProductRepository.java
@@ -1,15 +1,15 @@
 package com.platzi.market.domain.repository;
 
-import com.platzi.market.domain.Product;
+import com.platzi.market.domain.dto.Product;
 
 import java.util.List;
 import java.util.Optional;
 // Toma los métodos implementados son los orientados al dominio
 public interface ProductRepository {
     List<Product> getAll();
-    Optional<List<Product>> getByCategory(int categoryId);
-    Optional<List<Product>> getScarseProducts(int quantity);
-    Optional<Product> getProduct(int productID);
+    Optional<List<Product>> getByCategory(Long categoryId);
+    Optional<List<Product>> getScarseProducts(Long quantity);
+    Optional<Product> getProduct(Long productID);
     Product save(Product product);
-    void delete(int productId);
+    void delete(Long productId);
 }
