+++ src/main/java/com/platzi/market/persistence/entity/CompraProductoPK.java
@@ -7,23 +7,23 @@ import java.io.Serializable;
 @Embeddable
 public class CompraProductoPK implements Serializable {
     @Column(name = "id_compra")
-    private Integer idCompra;
+    private Long idCompra;
     @Column(name = "id_producto")
-    private Integer idProducto;
+    private Long idProducto;
 
-    public Integer getIdCompra() {
+    public Long getIdCompra() {
         return idCompra;
     }
 
-    public void setIdCompra(Integer idCompra) {
+    public void setIdCompra(Long idCompra) {
         this.idCompra = idCompra;
     }
 
-    public Integer getIdProducto() {
+    public Long getIdProducto() {
         return idProducto;
     }
 
-    public void setIdProducto(Integer idProducto) {
+    public void setIdProducto(Long idProducto) {
         this.idProducto = idProducto;
     }
 }
